core: first attempt at adding a common authentication infra, smack support.
Added a very simple authentication infra that lets you register
authentication backends and request authentication/access control
from a particular or any backend. The authentication API tries to
be a union of a SMACKish MAC (target / access-mode / id) and a
traditional authentication (target / id / authentication-token).
Hopefully it's enough to fill our needs without being too ugly or
unintuitive to use.

Currently there are two built-in authentication backends. One is
a straightforward SMACK backend, the other is an always-deny backend
which is installed to be the last backend in the chain.