build-sys: Build with json-c >= 0.11
The pkg-config file name changed with json-c 0.11, so configure checks
need to accordingly change. Also means we shouldn't try to use pkg-config
directly for the json_tokener_get_error() check.

As a result of this, it becomes apparent that murphy/common/json.h
implicitly depends on json.h (and thus so do all files that include it).
Fixing this would require moving out typedefs/macros linking mrp_json_*
to json_* from that header.