telephony: plugin arguments for resource set configuration
The resource set the plugin controls on behalf of the
telephony stack is now configurable using plugin arguments.
Currently the resource set zone, class and priority as well
as the playback and recording resource names, role,
optionality and shareability are configurable. The policies
are currently always set to relaxed. Adding configurability
for this using a plugin argument is now trivial if needed.

Change-Id: Ia80245c458014231fed1bb62cf3c660d091cd075