system-monitor: allow optional cgroup entries missing.
Some cgroup entries might be present or missing depending
on the active kernel configuration. Allow these entries
to be marked as such and don't consider it to be a fatal
error if these are missing.

At the moment, reading such an entry will return nil, and
writing to such an entry will silently fail but report
success. These semantics will need to be revisited...

Mark the swap-related entries in the memory controller
as potentially missing ones.

Change-Id: I0fd4013e6a4e6ff451271b9a5e3497cd1cf299a0