system-monitor: initial support for process watches.
This patch adds support for process tracking in the form of
process watches. Using a process watch on can get a notification
callback when an event of interest (fork, exec, exit, identity
change, etc.) happens in the tracked process.

Process watches relay on the low-level process watch interface
offered by the common Murphy libraries, which in turn rely in the
kernels process connector event support.

Change-Id: Ieeb2fc955a9b3f67054c28f5e1632ed236729f52