system-controller: try to get launching user for an app using ail and pid.
With the latest ail we need to retreive the launching user for an app
using the PID to get the category right. If the user  is not found this
way, we consider the default user ('app') as the launching user and fall
back to look up and use the UID of that instead.

Change-Id: Iaa1d9091598cd7b28dff0ee67e67d357325f152d
Signed-off-by: Baptiste DURAND <baptiste.durand@open.eurogiciel.org>